requirements.txt
numpy
scipy
librosa
soundfile
matplotlib
pandas
python-docx